Handle edge case in getSeekableStream method with maxStreamSize

+++ src/main/java/org/verapdf/io/InternalInputStream.java
@@ -280,6 +280,9 @@ public class InternalInputStream extends SeekableInputStream {
 	}
 
 	private static File createTempFile(byte[] alreadyRead, InputStream input, Integer maxStreamSize) throws IOException {
+		if (maxStreamSize != null && alreadyRead.length > maxStreamSize) {
+			throw new VeraPDFParserException("Maximum allowed stream size exceeded");
+		}
 		File tmpFile = File.createTempFile("tmp_pdf_file", ".pdf");
 		try (FileOutputStream output = new FileOutputStream(tmpFile)) {
 			output.write(alreadyRead);
